+++ packages/base/src/api/api.test.ts
@@ -3,7 +3,7 @@ import { createApi } from './api';
 
 jest.mock('axios');
 
-const apiUrl = 'http://platform-api.openzeppelin.com/';
+const apiUrl = 'http://defender-api.openzeppelin.com/';
 const key = 'key';
 const token = 'token';